try fix

+++ homework2/homework/models.py
@@ -31,7 +31,7 @@ class CNNClassifier(torch.nn.Module):
 
     def forward(self, x):
         x = self.network(x)
-        x = x.mean([2, 3])
+        x = float(x.mean([2, 3]))
         return self.classifier(x)[:,0]
